+++ include/openenclave/internal/posix/eventfd_ops.h
@@ -19,7 +19,7 @@ typedef struct _oe_eventfd_ops
 {
     oe_device_ops_t base;
 
-    oe_device_t* (*eventfd)(oe_device_t* dev, uint64_t initval, int flags);
+    oe_device_t* (*eventfd)(oe_device_t* dev, unsigned int initval, int flags);
 
 } oe_eventfd_ops_t;
